Refactor imports to utilize centralized search utilities for PlaceResult type, enhancing code consistency across date sets and favorites actions.
